+++ services/apps/entity_merging_worker/src/activities/members.ts
@@ -1,6 +1,8 @@
 import { WorkflowIdReusePolicy } from '@temporalio/workflow'
 
+import { updateActivities } from '@crowd/data-access-layer/src/activities/update'
 import { cleanupMemberAggregates } from '@crowd/data-access-layer/src/members/segments'
+import { IDbActivityCreateData } from '@crowd/data-access-layer/src/old/apps/data_sink_worker/repo/activity.data'
 import {
   cleanupMember,
   deleteMemberSegments,
@@ -8,7 +10,9 @@ import {
   getIdentitiesWithActivity,
   moveIdentityActivitiesToNewMember,
 } from '@crowd/data-access-layer/src/old/apps/entity_merging_worker'
-import { dbStoreQx } from '@crowd/data-access-layer/src/queryExecutor'
+import { figureOutNewOrgId } from '@crowd/data-access-layer/src/old/apps/profiles_worker'
+import { prepareMemberAffiliationsUpdate } from '@crowd/data-access-layer/src/old/apps/profiles_worker'
+import { dbStoreQx, pgpQx } from '@crowd/data-access-layer/src/queryExecutor'
 import { SearchSyncApiClient } from '@crowd/opensearch'
 import { RedisPubSubEmitter } from '@crowd/redis'
 import {
@@ -171,5 +175,75 @@ export async function notifyFrontendMemberUnmergeSuccessful(
       undefined,
       tenantId,
     ),
+  )
+}
+
+export async function finishMemberMergingUpdateActivities(memberId: string, newMemberId: string) {
+  const pgDb = svc.postgres.reader
+  const qDb = svc.questdbSQL
+  const queueClient = svc.queue
+
+  const qx = pgpQx(pgDb.connection())
+  const { orgCases } = await prepareMemberAffiliationsUpdate(qx, memberId)
+
+  await updateActivities(
+    qDb,
+    queueClient,
+    [
+      async () => ({ memberId: newMemberId }),
+      async (activity) => ({ organizationId: figureOutNewOrgId(activity, orgCases) }),
+    ],
+    `"memberId" = $(memberId)`,
+    {
+      memberId,
+    },
+  )
+}
+
+function moveByIdentities({
+  activity,
+  identities,
+  newMemberId,
+}: {
+  activity: IDbActivityCreateData
+  identities: IMemberIdentity[]
+  newMemberId: string
+}): Partial<IDbActivityCreateData> {
+  const { platform, username } = activity
+  const activityMatches = identities.some(
+    (i) =>
+      i.type === MemberIdentityType.USERNAME && i.platform === platform && i.value === username,
+  )
+
+  return activityMatches ? { memberId: newMemberId } : {}
+}
+
+export async function finishMemberUnmergingUpdateActivities({
+  memberId,
+  newMemberId,
+  identities,
+}: {
+  memberId: string
+  newMemberId: string
+  identities: IMemberIdentity[]
+}) {
+  const pgDb = svc.postgres.reader
+  const qDb = svc.questdbSQL
+  const queueClient = svc.queue
+
+  const qx = pgpQx(pgDb.connection())
+  const { orgCases } = await prepareMemberAffiliationsUpdate(qx, memberId)
+
+  await updateActivities(
+    qDb,
+    queueClient,
+    [
+      async (activity) => moveByIdentities({ activity, identities, newMemberId }),
+      async (activity) => ({
+        organizationId: figureOutNewOrgId(activity, orgCases),
+      }),
+    ],
+    `"memberId" = $(memberId)`,
+    { memberId },
   )
 }
